requirements.txt
setuptools~=51.0.0
streamlit~=0.87.0
streamlit-webrtc~=0.27.1
Pillow~=8.3.1
pika~=1.2.0
numpy~=1.19.4